refactor: split SjonskerdingarPage into components
